@@ -4,69 +4,71 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.0",
-    "@angular/cdk": "^13.1.3",
-    "@angular/common": "~13.1.0",
-    "@angular/compiler": "~13.1.0",
-    "@angular/core": "~13.1.0",
-    "@angular/forms": "~13.1.0",
-    "@angular/material": "^13.1.3",
-    "@angular/platform-browser": "~13.1.0",
-    "@angular/platform-browser-dynamic": "~13.1.0",
-    "@angular/router": "~13.1.0",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
-    "@nestjs/platform-express": "^8.0.0",
-    "@ngrx/component-store": "^13.0.2",
-    "@ngrx/effects": "^13.0.2",
-    "@ngrx/entity": "^13.0.2",
-    "@ngrx/router-store": "^13.0.2",
-    "@ngrx/store": "^13.0.2",
-    "@ngrx/store-devtools": "^13.0.2",
-    "@nrwl/angular": "13.4.6",
+    "@angular/animations": "14.2.4",
+    "@angular/cdk": "14.2.3",
+    "@angular/common": "14.2.4",
+    "@angular/compiler": "14.2.4",
+    "@angular/core": "14.2.4",
+    "@angular/forms": "14.2.4",
+    "@angular/material": "14.2.3",
+    "@angular/platform-browser": "14.2.4",
+    "@angular/platform-browser-dynamic": "14.2.4",
+    "@angular/router": "14.2.4",
+    "@nestjs/common": "9.1.2",
+    "@nestjs/core": "9.1.2",
+    "@nestjs/platform-express": "9.1.2",
+    "@ngrx/component-store": "14.0.2",
+    "@ngrx/effects": "14.0.2",
+    "@ngrx/entity": "14.0.2",
+    "@ngrx/router-store": "14.0.2",
+    "@ngrx/store": "14.0.2",
+    "@ngrx/store-devtools": "14.0.2",
+    "@nrwl/angular": "14.8.2",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.4.0",
+    "rxjs": "~7.5.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.0",
-    "@angular-eslint/eslint-plugin": "~13.0.1",
-    "@angular-eslint/eslint-plugin-template": "~13.0.1",
-    "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.1.0",
-    "@angular/compiler-cli": "~13.1.0",
-    "@angular/language-service": "~13.1.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "13.4.6",
-    "@nrwl/cypress": "13.4.6",
-    "@nrwl/eslint-plugin-nx": "13.4.6",
-    "@nrwl/jest": "13.4.6",
-    "@nrwl/linter": "13.4.6",
-    "@nrwl/nest": "13.4.6",
-    "@nrwl/node": "13.4.6",
-    "@nrwl/tao": "13.4.6",
-    "@nrwl/workspace": "13.4.6",
-    "@types/jest": "27.0.2",
-    "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.3.0",
-    "@typescript-eslint/parser": "~5.3.0",
+    "@angular-devkit/build-angular": "14.2.4",
+    "@angular-eslint/eslint-plugin": "14.0.4",
+    "@angular-eslint/eslint-plugin-template": "14.0.4",
+    "@angular-eslint/template-parser": "14.0.4",
+    "@angular/cli": "~14.2.0",
+    "@angular/compiler-cli": "14.2.4",
+    "@angular/language-service": "14.2.4",
+    "@nestjs/schematics": "9.0.3",
+    "@nestjs/testing": "9.1.2",
+    "@nrwl/cli": "14.8.2",
+    "@nrwl/cypress": "14.8.2",
+    "@nrwl/eslint-plugin-nx": "14.8.2",
+    "@nrwl/jest": "14.8.2",
+    "@nrwl/linter": "14.8.2",
+    "@nrwl/nest": "14.8.2",
+    "@nrwl/node": "14.8.2",
+    "@nrwl/workspace": "14.8.2",
+    "@types/jest": "28.1.8",
+    "@types/node": "18.7.1",
+    "@typescript-eslint/eslint-plugin": "5.38.1",
+    "@typescript-eslint/parser": "5.38.1",
     "cypress": "^9.1.0",
-    "eslint": "8.2.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
-    "prettier": "^2.3.1",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.4.3"
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.2",
+    "prettier": "2.7.1",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4",
+    "nx": "14.8.2"
   }
 }
